scripts/requirements.txt
googlemaps
openai
supabase
python-dotenv
tqdm
tenacity